Move UnresolvedConfigurationCondition to com.oracle.svm.configure
