+++ substratevm/src/com.oracle.svm.configure/src/com/oracle/svm/configure/ConditionalElement.java
@@ -28,8 +28,6 @@ package com.oracle.svm.configure;
 import java.util.Comparator;
 import java.util.function.Function;
 
-import org.graalvm.nativeimage.impl.UnresolvedConfigurationCondition;
-
 public record ConditionalElement<T>(UnresolvedConfigurationCondition condition, T element) {
 
     public static <T extends Comparable<T>> Comparator<ConditionalElement<T>> comparator() {
